requirements.txt
streamlit==1.30.0
google-generativeai==0.3.1
python-dotenv==1.0.0
manim==0.17.3
pillow==10.0.1